Refactor to go layout with internal folder
