+++ code-example/simple-service/internal/api/handlers/ping.go
